fix(blockchain): remove duplicate sync status tracker
